requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0
PyGithub>=1.59.0
python-dotenv>=1.0.0
requests>=2.31.0